Moves one hot to array ops
